requirements.txt
google-cloud-vision>=0.38.0
logbook>=1.4.3
python-dotenv>=0.10.3
python-telegram-bot==12.0.0b1
requests>=2.22.0
slackclient>=2.0.1
textblob>=0.15.3
